Fix more issues post-merge
