+++ frameioclient/lib/service.py
@@ -1,7 +1,8 @@
+from ..client import FrameioClient
 from ..lib.bandwidth import NetworkBandwidth
 
 class Service(object):
-    def __init__(self, client):
+    def __init__(self, client: FrameioClient):
         self.client = client
         self.concurrency = 10
         self.bandwidth = NetworkBandwidth()
