--- frameioclient/service/service.py
@@ -1,5 +1,0 @@
-from ..client import FrameioClient
-
-class Service(object):
-    def __init__(self, client: FrameioClient):
-        self.client = client
