+++ frameioclient/services/projects.py
@@ -1,4 +1,5 @@
 from ..lib.service import Service
+from .helpers import FrameioHelpers
 
 class Project(Service):
   def create(self, team_id, **kwargs):
@@ -34,7 +35,48 @@ class Project(Service):
     """
     endpoint = '/projects/{}'.format(project_id)
     return self.client._api_call('get', endpoint)
-  
+
+  def tree(self, project_id, slim):
+    """
+    Fetch a tree representation of all files/folders in a project.
+
+    :Args:
+      project_id (string): The project's id
+      slim (bool): If true, fetch only the minimum information for the following:
+        filename,
+        filesize,
+        thumbnail,
+        creator_id,
+        inserted_at (date created),
+        path (represented like a filesystem)
+
+      Example::
+        client.projects.get(
+          project_id="123",
+          slim=True
+        )
+    """
+    # endpoint = "/projects/{}/tree?depth=20&drop_includes=a.transcode_statuses,a.transcodes,a.source,a.checksums&only_fields=a.name,a.filesize,u.name,a.item_count,a.creator_id,a.inserted_at,a.uploaded_at".format(project_id)
+    # return self.client._api_call('get', endpoint)
+
+    return FrameioHelpers(self.client).build_project_tree(project_id, slim)
+
+  def download(self, project_id, destination_directory='downloads'):
+    """
+    Download the provided project to disk.
+
+    :Args:
+      project_id (uuid): The project's id.
+      destination_directory (string): Directory on disk that you want to download the project to.
+
+      Example::
+        client.projects.download(
+          project_id="123",
+          destination_directory="./downloads"
+        )
+    """
+    return FrameioHelpers(self.client).download_project(project_id, destination=destination_directory)
+
   def get_collaborators(self, project_id, **kwargs):
     """
     Get collaborators for a project
